arm: KVM: Invalidate BTB on guest exit
In order to avoid aliasing attacks against the branch predictor,
let's invalidate the BTB on guest exit. This is made complicated
by the fact that we cannot take a branch before invalidating the
BTB.

Another thing is that we perform the invalidation on all
implementations, no matter if they are affected or not.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit 8f4c8defa0cfc2c8d53d095eeda6c6ad1440201b)
Signed-off-by: Alex Shi <alex.shi@linaro.org>

Conflicts:
	patch on old hyp_hvc: in arch/arm/kvm/hyp/hyp-entry.S