PCI/DPC: Consolidate RP PIO get/print functions
Previously we defined a structure for RP PIO log information, allocated it
on the stack, called one function to fill it in from DPC registers, and
called another to print it out.

Simplify this by dropping the structure and printing the error information
as soon as we read it from the DPC capability.  This way we don't need a
structure, there's one function instead of four, and everything we do with
the register information is in one place instead of being split between
functions.

No functional change intended.

Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Keith Busch <keith.busch@intel.com>